Nuevas apis
Nuevas apis

+++ bjitgad/src/main/java/com/jitgad/bjitgad/AllApis/Activitiestyperesource.java
@@ -58,7 +58,7 @@ public class Activitiestyperesource {
      */
     @GET
     @Produces(MediaType.APPLICATION_JSON)
-    public Response getActivitiestype(){
+    public Response getActivitiestype() {
         String data = atC.selectActivitiestype();
         String responseJson = Rapi.Response("Ocurrió un error", false, data);
         try {
@@ -88,16 +88,20 @@ public class Activitiestyperesource {
             String Authorization = headers.getHeaderString("Authorization");
             Authorization = Authorization == null ? "" : Authorization;
             System.out.println("Authorization: " + Authorization);
-            Object[] Permt = AuC.VToken(Authorization);
-            if (Permt[0].equals(true)) {
-                responseCountingPage = atC.CountingPageActivitiesType();
-                if (responseJson.equals("{}")) {
-                    responseJson = Rapi.AdminResponse("Información no encontrada", responseCountingPage, false, responseJson);
+            if (!Authorization.isEmpty()) {
+                Object[] Permt = AuC.VToken(Authorization);
+                if (Permt[0].equals(true)) {
+                    responseCountingPage = atC.CountingPageActivitiesType();
+                    if (responseJson.equals("{}")) {
+                        responseJson = Rapi.AdminResponse("Información no encontrada", responseCountingPage, false, responseJson);
+                    } else {
+                        responseJson = Rapi.AdminResponse("Datos retornados correctamente", responseCountingPage, true, responseJson);
+                    }
                 } else {
-                    responseJson = Rapi.AdminResponse("Datos retornados correctamente", responseCountingPage, true, responseJson);
+                    responseJson = Rapi.AdminResponse(String.valueOf(Permt[1]), responseCountingPage, false, responseJson);
                 }
             } else {
-                responseJson = Rapi.AdminResponse(String.valueOf(Permt[1]), responseCountingPage, false, responseJson);
+                responseJson = Rapi.AdminResponse("Token vacio", responseCountingPage, false, responseJson);
             }
         } catch (Exception e) {
             responseJson = Rapi.AdminResponse(e.getMessage(), responseCountingPage, false, responseJson);
@@ -116,21 +120,25 @@ public class Activitiestyperesource {
     public Response getactivitiesbyid(@Context HttpHeaders headers, @QueryParam("activityid") String activityid) {
         String data = "{}";
         String responseJson = Rapi.Response("Ocurrió un error", false, data);
-        //TOKENS
-        String Authorization = headers.getHeaderString("Authorization");
-        Authorization = Authorization == null ? "" : Authorization;
-        System.out.println("Authorization: " + Authorization);
-        Object[] Permt = AuC.VToken(Authorization);
         try {
-            if (Permt[0].equals(true)) {
-                data = atC.selectactivitiesbyid(activityid);
-                if (data.equals("{}")) {
-                    responseJson = Rapi.Response("Información no encontrada", false, data);
+            //TOKENS
+            String Authorization = headers.getHeaderString("Authorization");
+            Authorization = Authorization == null ? "" : Authorization;
+            System.out.println("Authorization: " + Authorization);
+            if (!Authorization.isEmpty()) {
+                Object[] Permt = AuC.VToken(Authorization);
+                if (Permt[0].equals(true)) {
+                    data = atC.selectactivitiesbyid(activityid);
+                    if (data.equals("{}")) {
+                        responseJson = Rapi.Response("Información no encontrada", false, data);
+                    } else {
+                        responseJson = Rapi.Response("Datos retornados correctamente", true, data);
+                    }
                 } else {
-                    responseJson = Rapi.Response("Datos retornados correctamente", true, data);
+                    responseJson = Rapi.Response(String.valueOf(Permt[1]), false, data);
                 }
             } else {
-                responseJson = Rapi.Response(String.valueOf(Permt[1]), false, data);
+                responseJson = Rapi.Response("Tokén vacio", true, data);
             }
 
         } catch (Exception e) {
@@ -164,7 +172,7 @@ public class Activitiestyperesource {
     @Path("/postActivitiesType")
     @Consumes(MediaType.APPLICATION_JSON)
     public Response PostActivitiesType(@Context HttpHeaders headers, String data) {
-        String responseJson = "{\"status\":\"poken:" + data + "\"}";
+        String responseJson = Rapi.Response("Ocurrió un error", false, data);
         System.out.println("Ingresando PostActivitiesType...");
         JsonObject Jso = Methods.stringToJSON(data);
         try {
@@ -174,27 +182,29 @@ public class Activitiestyperesource {
                 String Authorization = headers.getHeaderString("Authorization");
                 Authorization = Authorization == null ? "" : Authorization;
                 System.out.println("Authorization: " + Authorization);
-                Object[] Permt = AuC.VToken(Authorization);
-                if (Permt[0].equals(true)) {
-                    responseatC = atC.InsertActivitiesTypeC(
-                            Methods.JsonToString(Jso.getAsJsonObject(), "name", ""),
-                            Methods.JsonToString(Jso.getAsJsonObject(), "image", ""),
-                            // Methods.JsonToString(Jso.getAsJsonObject(), "creationdate", ""),
-                            //  Methods.JsonToString(Jso.getAsJsonObject(), "updatedate", ""),
-                            Methods.JsonToString(Jso.getAsJsonObject(), "state", ""));
-                    if (responseatC[0].equals(true)) {
-                        responseJson = Rapi.Response(String.valueOf(responseatC[1]), Boolean.parseBoolean(responseatC[0].toString()), data);
+                if (!Authorization.isEmpty()) {
+                    Object[] Permt = AuC.VToken(Authorization);
+                    if (Permt[0].equals(true)) {
+                        responseatC = atC.InsertActivitiesTypeC(
+                                Methods.JsonToString(Jso.getAsJsonObject(), "name", ""),
+                                Methods.JsonToString(Jso.getAsJsonObject(), "image", ""),
+                                Methods.JsonToString(Jso.getAsJsonObject(), "state", ""));
+                        if (responseatC[0].equals(true)) {
+                            responseJson = Rapi.Response(String.valueOf(responseatC[1]), Boolean.parseBoolean(responseatC[0].toString()), String.valueOf(responseatC[2]));
+                        } else {
+                            responseJson = Rapi.Response(String.valueOf(responseatC[1]), Boolean.parseBoolean(responseatC[0].toString()), String.valueOf(responseatC[2]));
+                        }
                     } else {
-                        responseJson = Rapi.Response(String.valueOf(responseatC[1]), Boolean.parseBoolean(responseatC[0].toString()), data);
+                        responseJson = Rapi.Response(String.valueOf(Permt[1]), false, "{}");
                     }
                 } else {
-                    responseJson = Rapi.Response(String.valueOf(Permt[1]), false, data);
+                    responseJson = Rapi.Response("Tokén vacio", true, data);
                 }
             } else {
-                responseJson = Rapi.Response("Información no encontrada", false, data);
+                responseJson = Rapi.Response("Información no encontrada", false, "{}");
             }
         } catch (Exception e) {
-            responseJson = Rapi.Response(e.getMessage(), false, data);
+            responseJson = Rapi.Response(e.getMessage(), false, "{}");
         }
         return Response.ok(responseJson)
                 .header("Access-Control-Allow-Methods", "POST, GET, PUT, UPDATE, OPTIONS")
@@ -206,7 +216,7 @@ public class Activitiestyperesource {
     @PUT
     @Path("/PutActivitiesType")
     public Response PutActivitiesType(@Context HttpHeaders headers, String data) {
-        String responseJson = "{\"status\":\"poken:" + data + "\"}";
+        String responseJson = Rapi.Response("Ocurrió un error", false, data);
         System.out.println("Ingresando PutActivitiesType...");
         JsonObject Jso = Methods.stringToJSON(data);
         try {
@@ -216,26 +226,30 @@ public class Activitiestyperesource {
                 String Authorization = headers.getHeaderString("Authorization");
                 Authorization = Authorization == null ? "" : Authorization;
                 System.out.println("Authorization: " + Authorization);
-                Object[] Permt = AuC.VToken(Authorization);
-                if (Permt[0].equals(true)) {
-                    responseatC = atC.UpdateActivitiesTypeC(
-                            Methods.JsonToInteger(Jso.getAsJsonObject(), "idactivitiestype", 0),
-                            Methods.JsonToString(Jso.getAsJsonObject(), "name", ""),
-                            Methods.JsonToString(Jso.getAsJsonObject(), "image", ""),
-                            Methods.JsonToString(Jso.getAsJsonObject(), "state", ""));
-                    if (responseatC[0].equals(true)) {
-                        responseJson = Rapi.Response(String.valueOf(responseatC[1]), Boolean.parseBoolean(responseatC[0].toString()), String.valueOf(responseatC[2]));
+                if (!Authorization.isEmpty()) {
+                    Object[] Permt = AuC.VToken(Authorization);
+                    if (Permt[0].equals(true)) {
+                        responseatC = atC.UpdateActivitiesTypeC(
+                                Methods.JsonToInteger(Jso.getAsJsonObject(), "idactivitiestype", 0),
+                                Methods.JsonToString(Jso.getAsJsonObject(), "name", ""),
+                                Methods.JsonToString(Jso.getAsJsonObject(), "image", ""),
+                                Methods.JsonToString(Jso.getAsJsonObject(), "state", ""));
+                        if (responseatC[0].equals(true)) {
+                            responseJson = Rapi.Response(String.valueOf(responseatC[1]), Boolean.parseBoolean(responseatC[0].toString()), "{}");
+                        } else {
+                            responseJson = Rapi.Response(String.valueOf(responseatC[1]), Boolean.parseBoolean(responseatC[0].toString()), "{}");
+                        }
                     } else {
-                        responseJson = Rapi.Response(String.valueOf(responseatC[1]), Boolean.parseBoolean(responseatC[0].toString()), String.valueOf(responseatC[2]));
+                        responseJson = Rapi.Response(String.valueOf(Permt[1]), false, "{}");
                     }
                 } else {
-                    responseJson = Rapi.Response(String.valueOf(Permt[1]), false, "");
+                    responseJson = Rapi.Response("Tokén vacio", true, data);
                 }
             } else {
-                responseJson = Rapi.Response("Información no encontrada", false, "");
+                responseJson = Rapi.Response("Información no encontrada", false, "{}");
             }
         } catch (Exception e) {
-            responseJson = Rapi.Response(e.getMessage(), false, "");
+            responseJson = Rapi.Response(e.getMessage(), false, "{}");
         }
         return Response.ok(responseJson)
                 .header("Access-Control-Allow-Methods", "POST, GET, PUT, UPDATE, OPTIONS")
@@ -247,7 +261,7 @@ public class Activitiestyperesource {
     @DELETE
     @Path("/DeleteActivitiesType")
     public Response DeleteActivitiesType(@Context HttpHeaders headers, String data) {
-        String responseJson = "{\"status\":\"poken:" + data + "\"}";
+        String responseJson = Rapi.Response("Ocurrió un error", false, data);
         System.out.println("Ingresando DeleteActivitiesType...");
         JsonObject Jso = Methods.stringToJSON(data);
         System.out.println(responseJson);
@@ -258,24 +272,26 @@ public class Activitiestyperesource {
                 String Authorization = headers.getHeaderString("Authorization");
                 Authorization = Authorization == null ? "" : Authorization;
                 System.out.println("Authorization: " + Authorization);
-                Object[] Permt = AuC.VToken(Authorization);
-              
-                if (Permt[2].equals("Administrador")) {
-                    if (Permt[0].equals(true)) {
-                        responseatC = atC.DeleteActividadestype(
-                                Methods.JsonToInteger(Jso.getAsJsonObject(), "idactivitiestype", 0));
-                        if (responseatC[0].equals(true)) {
-                            responseJson = Rapi.Response(String.valueOf(responseatC[1]), Boolean.parseBoolean(responseatC[0].toString()), data);
+                if (!Authorization.isEmpty()) {
+                    Object[] Permt = AuC.VToken(Authorization);
+                    if (Permt[2].equals("Administrador")) {
+                        if (Permt[0].equals(true)) {
+                            responseatC = atC.DeleteActividadestype(
+                                    Methods.JsonToInteger(Jso.getAsJsonObject(), "idactivitiestype", 0));
+                            if (responseatC[0].equals(true)) {
+                                responseJson = Rapi.Response(String.valueOf(responseatC[1]), Boolean.parseBoolean(responseatC[0].toString()), data);
+                            } else {
+                                responseJson = Rapi.Response(String.valueOf(responseatC[1]), Boolean.parseBoolean(responseatC[0].toString()), data);
+                            }
                         } else {
-                            responseJson = Rapi.Response(String.valueOf(responseatC[1]), Boolean.parseBoolean(responseatC[0].toString()), data);
+                            responseJson = Rapi.Response(String.valueOf(Permt[1]), false, data);
                         }
                     } else {
-                        responseJson = Rapi.Response(String.valueOf(Permt[1]), false, data);
+                        responseJson = Rapi.Response("Usuario sin privilegios para realizar esta actividad", false, data);
                     }
-                }else{
-                    responseJson = Rapi.Response("Usuario sin privilegios para realizar esta actividad", false, data);
+                } else {
+                    responseJson = Rapi.Response("Tokén vacio", true, data);
                 }
-
             } else {
                 responseJson = Rapi.Response("Información no encontrada", false, data);
             }
@@ -287,8 +303,7 @@ public class Activitiestyperesource {
                 .header("Access-Control-Allow-Headers", "Content-Type, Accept, X-Requested-with")
                 .build();
     }
-    
-    
+
     /*
      add activities
      */
@@ -305,9 +320,9 @@ public class Activitiestyperesource {
                 String name = Methods.JsonToString(Jso.getAsJsonObject(), "name", "");
                 String base64 = Methods.JsonToString(Jso.getAsJsonObject(), "base64", "");
                 FileController fc = new FileController();
-                Object[] CreateFile = fc.createfile(base64,"Activities",name);
-                responseJson = Rapi.Response("Imagen creada con éxito", Boolean.parseBoolean(CreateFile[0].toString()), String.valueOf(CreateFile[1].toString()+ "/"+ name+ "/"+CreateFile[2].toString()));
-                
+                Object[] CreateFile = fc.createfile(base64, "Activities", name);
+                responseJson = Rapi.Response("Imagen creada con éxito", Boolean.parseBoolean(CreateFile[0].toString()), String.valueOf(CreateFile[1].toString() + "/" + name + "/" + CreateFile[2].toString()));
+
             } else {
                 responseJson = Rapi.Response("Información no encontrada", false, data);
             }
